Fixed linting errors and scrollbar
