+++ app/src/components/SelectPages/SelectPages.module.css
@@ -2,18 +2,20 @@ h2 {
     text-align: left;
     position: relative;
     top: 50px;
-    left: 1vw;
+    left: 1%;
     margin: 0px;
     font-family: Oswald, sans-serif;
     font-size: 48px;
+    width: 99%;
 }
 
 .container {
     position: relative;
-    left: 5vw;
+    left: 5%;
     margin-top: 50px;
     font-family: Oswald, sans-serif;
     font-size: 36px;
+    width: 95%;
 }
 
 .pageLabel {
